+++ vs/dosbox.vcxproj
@@ -277,6 +277,7 @@
     <ClInclude Include="..\include\dosbox.h" />
     <ClInclude Include="..\include\dos_inc.h" />
     <ClInclude Include="..\include\dos_system.h" />
+    <ClInclude Include="..\include\drives.h" />
     <ClInclude Include="..\include\fpu.h" />
     <ClInclude Include="..\include\hardware.h" />
     <ClInclude Include="..\include\inout.h" />
@@ -339,7 +340,6 @@
     <ClInclude Include="..\src\dos\cdrom.h" />
     <ClInclude Include="..\src\dos\dev_con.h" />
     <ClInclude Include="..\src\dos\dos_mscdex.h" />
-    <ClInclude Include="..\src\dos\drives.h" />
     <ClInclude Include="..\src\dos\Ntddcdrm.h" />
     <ClInclude Include="..\src\dos\Ntddscsi.h" />
     <ClInclude Include="..\src\dos\Ntddstor.h" />
